python-fd: merge mariabackup plugin files
