+++ debian/bareos-filedaemon-mariabackup-python-plugin.install.in
@@ -1,2 +1,1 @@
 @plugindir@/bareos-fd-mariabackup.py*
-@plugindir@/BareosFdPluginMariabackup.py*
